ensure you can click anywhere outside modal to close
before this change, if your click was wider than the <body> tag, you'd
also miss the overlay so you wouldn't close the modal.

This changes the overlay to fill the open window, escaping the bounds of
the <body> tag.  The modal then needs some styling to ensure it fits
within the body.